Work on refactoring to be more OOPy and extensible
* Create custom transformers directory to keep implementation-specific data processing scripts
* Reorganize code to be more conventional and utilize modules
* Move ILAO code to custom transformer from support module
